+++ Atm_comparator.hpp
@@ -14,7 +14,7 @@
 #define STATE_TYPE state_t
 #endif
 
-typedef void ( *atm_comparator_cb_t )( int16_t idx, int v, int up, int idx_threshold, int v_threshold );
+typedef void ( *atm_comparator_cb_t )( int idx, int v, int up, int idx_threshold, int v_threshold );
 
 class Atm_comparator : public MACHINE {
  public:
@@ -45,11 +45,11 @@ class Atm_comparator : public MACHINE {
   Atm_comparator& begin( int attached_pin, int sampleRate = 50 );
   Atm_comparator& threshold( uint16_t* v, uint16_t size, bool catchUp = false );
   Atm_comparator& average( uint16_t* v, uint16_t size );
-  Atm_comparator& onUp( atm_comparator_cb_t callback, int16_t idx = 0 );
+  Atm_comparator& onUp( atm_comparator_cb_t callback, int idx = 0 );
   Atm_comparator& onUp( Machine& machine, int event = 0 );
   Atm_comparator& onUp( TinyMachine& machine, int event = 0 );
   Atm_comparator& onUp( const char* label, int event = 0 );
-  Atm_comparator& onDown( atm_comparator_cb_t callback, int16_t idx = 0 );
+  Atm_comparator& onDown( atm_comparator_cb_t callback, int idx = 0 );
   Atm_comparator& onDown( Machine& machine, int event = 0 );
   Atm_comparator& onDown( TinyMachine& machine, int event = 0 );
   Atm_comparator& onDown( const char* label, int event = 0 );
